Refactor to use custom firebase hook, scrollTo in useEffect mount, and useRef hook for easier testability
