+++ src/pages/Artist.tsx
@@ -1,9 +1,6 @@
-import React, { useEffect, useState, useRef } from 'react';
-import { useHistory } from 'react-router-dom';
-import firebaseService from '../util/services/firebase.service';
+import React, { useEffect } from 'react';
 import Layout from '../components/Layout';
 import Chatbox from '../components/Chatbox';
-import { useAuthContext } from '../util/providers/AuthProvider';
 import styled from 'styled-components';
 
 export const ArtistPageContainer = styled.div`
@@ -25,64 +22,19 @@ export const ArtistImg = styled.img`
   object-position: 90%;
 `;
 
-interface Chat {
-  avatar: string;
-  date: number;
-  msg: string;
-  sender: string;
-}
-
-export default function Artist() {
-  const history = useHistory();
-  const { currentUser } = useAuthContext();
-  const [chats, addChats] = useState<Chat[]>([]);
-  const { actor, name, avatar } = currentUser;
-  const chatlist = useRef<HTMLUListElement | null>(null);
-
+const Artist = () => {
   useEffect(() => {
-    firebaseService
-      .collection('members')
-      .where('user', '==', actor)
-      .get()
-      .then((querySnapshot) => {
-        if (querySnapshot.empty) {
-          history.push('/landing');
-          return;
-        }
-      });
-  }, [actor, history]);
-
-  useEffect(() => {
-    const unsubscribe = firebaseService
-      .collection('chats')
-      .orderBy('date')
-      .onSnapshot((snapshot) => {
-        if (snapshot.size) {
-          const firebaseChats: Chat[] = [];
-          snapshot.forEach((doc) => {
-            firebaseChats.push(doc.data() as Chat);
-          });
-          addChats(firebaseChats);
-          if (chatlist && chatlist.current) {
-            chatlist.current.scrollTop = chatlist.current.scrollHeight;
-          }
-        }
-      });
-
-    return () => unsubscribe();
+    window.scrollTo(0, 0);
   }, []);
 
   return (
     <Layout>
       <ArtistPageContainer>
-        <Chatbox
-          chats={chats}
-          sender={name}
-          avatar={avatar}
-          chatlist={chatlist}
-        />
+        <Chatbox />
         <ArtistImg src="/girl.png" alt="artist" />
       </ArtistPageContainer>
     </Layout>
   );
-}
+};
+
+export default Artist;
